[FABC-774] Fix inability to set environment variable
Value for environment variable `FABRIC_CA_SERVER_COMPATIBILITY_MODE_V1_3`
would always be set to true, even if the environment variable was explicitly
set to another value (e.g. false).

If environment variable is not set, the default will be true. If a
value has been explicitly set using the env var, then the set will be used
and the default will be overridden.

Change-Id: Icf5b3d2d543731f72084771fb083a5af54a78ee1
Signed-off-by: Saad Karim <skarim@us.ibm.com>
Signed-off-by: 陈桂军 <chengj16@bngrp.com>