Fix --csr.hosts flag for client and server
The docs currently state the --csr.hosts
takes a space-separated list but it should
actually be a comma-separate list.

Updated the help for the command flags and
regenerated the docs as well.

Also adding logic to trim spaces and
modified the function which checks SANs
to use built-in fields of
x509.Certificate.

FABC-831 #done

Change-Id: I41ebeeb98997fe5ad1672e62076214ad6a9d7c4a
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>
Signed-off-by: 陈桂军 <chengj16@bngrp.com>