Replace LabelHelp with info in doc template
The label "help" that was added to the metrics options really didn't do
much to clarify what the labels contained and it had a side effect of
pulling in code from a branch of fabric that is newer than the CA
release.

This change replaces use of LabelHelp with additional information in the
metrics template and switches metrics back to the release-1.4 branch of
fabric.

Signed-off-by: Matthew Sykes <sykesmat@us.ibm.com>
Signed-off-by: 陈桂军 <chengj16@bngrp.com>